refactor(Profile): RHICOMPL-3402 extract retrieval rules and groups
