+++ app/models/concerns/profile_fields.rb
@@ -38,20 +38,6 @@ module ProfileFields
       parent_profile_id.blank?
     end
 
-    def rules_and_rule_groups
-      conflicts = relationships_for('conflicts')
-      requires = relationships_for('requires')
-      arranged_rule_groups = rule_groups.includes(:rules).arrange_serializable do |parent, children|
-        { 'rule_group' => parent, 'group_children' => children,
-          'rule_children' => parent.rules_with_relationships(requires, conflicts),
-          'requires' => requires[parent], 'conflicts' => conflicts[parent] }
-      end
-
-      arranged_rule_groups.concat(rules.without_rule_group_parent.map do |pr|
-        { 'rule' => pr, 'requires' => requires[pr], 'conflicts' => conflicts[pr] }
-      end)
-    end
-
     private
 
     def bm_versions
@@ -60,15 +46,6 @@ module ProfileFields
         ref_id: ref_id,
         upstream: false
       ).joins(:benchmark).pluck('benchmarks.version')
-    end
-
-    def relationships_for(relationship)
-      RuleGroupRelationship.with_relationships(rules, relationship).or(
-        RuleGroupRelationship.with_relationships(rule_groups, relationship)
-      ).each_with_object({}) do |rgr, relationships|
-        relationships[rgr.left] ||= []
-        relationships[rgr.left] << rgr.right
-      end
     end
   end
 end
